pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dell-powerstore-mcp-server"
version = "1.0.0"
description = "MCP Server for Dell PowerStore with automatic OpenAPI tool generation - Credential-free mode"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "sachdev27", email = "opensource@sachdev27.com"}
]
maintainers = [
    {name = "sachdev27", email = "opensource@sachdev27.com"}
]
keywords = [
    "mcp",
    "model-context-protocol",
    "dell",
    "powerstore",
    "storage",
    "openapi",
    "ai",
    "llm",
    "claude",
    "automation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "starlette>=0.36.0",
    "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/sachdev27/dell-powerstore-mcp-server"
Documentation = "https://github.com/sachdev27/dell-powerstore-mcp-server#readme"
Repository = "https://github.com/sachdev27/dell-powerstore-mcp-server.git"
Issues = "https://github.com/sachdev27/dell-powerstore-mcp-server/issues"
Changelog = "https://github.com/sachdev27/dell-powerstore-mcp-server/blob/main/CHANGELOG.md"

[project.scripts]
powerstore-mcp = "powerstore_mcp.main:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["powerstore_mcp*"]

[tool.setuptools.package-data]
powerstore_mcp = ["py.typed"]

# ============================================================================
# Black - Code Formatter
# ============================================================================
[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# ============================================================================
# Ruff - Linter
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["powerstore_mcp"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]

# ============================================================================
# MyPy - Type Checker
# ============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
strict_equality = true
strict_concatenate = true
check_untyped_defs = true
disallow_untyped_defs = false
disallow_incomplete_defs = true
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "httpx.*",
    "yaml.*",
]
ignore_missing_imports = true

# ============================================================================
# Pytest - Testing
# ============================================================================
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ============================================================================
# Coverage - Test Coverage
# ============================================================================
[tool.coverage.run]
source = ["powerstore_mcp"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
show_missing = true
skip_covered = true
fail_under = 70
